requirements.txt
pymongo
fastapi
uvicorn
pydantic
python-dotenv